pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartchat-automation"
version = "0.1.0"
description = "Nowoczesny komunikator webowy z automatyzacją konwersacji."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Codex", email = "codex@example.com" }
]
dependencies = [
  "aiosqlite>=0.20.0",
  "alembic>=1.14.0",
  "asyncpg>=0.29.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.4",
  "orjson>=3.10.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.12",
  "redis>=5.2.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.7.3"
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"
